Create coppermine_to_dirstruct_skipper.py
Since I had a couple of files missing in my source folder, I created this modified version of the script.

In this modified version of the script, I added a try-except block around the shutil.copy2 function, which catches any exceptions that may be raised if the file cannot be copied. I also added a check to see if the source file exists, and if it doesn't, the script skips it and prints a message to the console which files it skipped.